Segregate the entire versions stack
