+++ src/app.rs
@@ -17,8 +17,8 @@ pub fn rocket_from_config(figment: Figment) -> Rocket<Build> {
                 routes::v1::configs_routes::get_configs,
                 routes::v1::configs_routes::get_config,
                 routes::v1::configs_routes::delete_config,
-                routes::v1::configs_routes::create_version,
-                routes::v1::configs_routes::get_versions,
+                routes::v1::versions_routes::create_version,
+                routes::v1::versions_routes::get_versions,
             ],
         )
         .register(
