+++ src/repos/mod.rs
@@ -1,2 +1,3 @@
 pub mod configs_repo;
 pub mod domains_repo;
+pub mod versions_repo;
