+++ src/routes/v1/mod.rs
@@ -1,4 +1,5 @@
 pub mod configs_routes;
 pub mod domains_routes;
+pub mod versions_routes;
 pub mod dtos;
 pub mod errors;
