+++ src/services/mod.rs
@@ -1,2 +1,3 @@
 pub mod configs_service;
 pub mod domains_service;
+pub mod versions_service;
